+++ src/main/kotlin/kravis/GGPlot2.kt
@@ -336,9 +336,9 @@ object OrderUtils {
      */
     fun reorder(f: String, orderAttribute: String, orderFun: OrderFun = OrderFun.mean, ascending: Boolean = true): String {
         return (if (ascending) {
-            "fct_reorder($f, $orderAttribute, fun=$orderFun)"
+            "fct_reorder($f, $orderAttribute, .fun=$orderFun)"
         } else {
-            "fct_rev(fct_reorder($f, $orderAttribute, fun=$orderFun))"
+            "fct_rev(fct_reorder($f, $orderAttribute, .fun=$orderFun))"
         }).asRExpression
     }