+++ src/main/kotlin/kravis/device/JupyterDevice.kt
@@ -1,6 +1,5 @@
 package kravis.device
 
-import jupyter.kotlin.MimeTypedResult
 import krangl.irisData
 import kravis.Aesthetic
 import kravis.GGPlot
@@ -8,6 +7,7 @@ import kravis.SessionPrefs
 import kravis.geomPoint
 import kravis.nshelper.plot
 import kravis.render.PlotFormat
+import org.jetbrains.kotlinx.jupyter.api.MimeTypedResult
 import java.awt.Dimension
 import java.awt.image.BufferedImage
 import java.io.ByteArrayOutputStream